servohack: Only cycle USB Power when usercode drive is removed
If we cycle USB power at the end of a usercode program whilst the USB drive
is still inserted, it kicks the USB drive out and in, which astdiskd then sees
as a new drive insertion and we end up with the robot in a restart loop.

This change means that we cannot restart the code without pulling the USB,
although this is always going to have to be the case until PLOD is fixed
as we cannot determine the order that the user is going to plug in USB devices.

For example, if the user used a "dumb hub", (no per-port power switching) and
plugged in both the usercode drive and servo board, we would be unable to
restart the servo board without also power cycling the usercode drive too.